refactor: split parser methods as extensions
