+++ src/main/kotlin/drift/cli/DriftRunner.kt
@@ -1,4 +1,4 @@
-package drift
+package drift.cli
 
 import drift.ast.eval
 import drift.check.SymbolCollector
@@ -7,7 +7,6 @@ import drift.parser.Parser
 import drift.parser.lex
 import drift.runtime.*
 import java.io.File
-import kotlin.system.exitProcess
 
 fun main(args: Array<String>) {
     if (args.isEmpty()) {
